fix: code org, navigation and debounce
